requirements.txt
..
asgiref==3.7.2
blinker==1.7.0
build==1.0.3
certifi
charset-normalizer
click==8.1.7
DateTime==5.4
Django==4.2.8
docutils==0.20.1
Flask==3.0.0
Flask-DebugToolbar==0.14.1
Flask-Login==0.6.3
flask-migrate
Flask-SQLAlchemy==3.1.1
flask_profiler
idna
importlib-metadata==7.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonpickle==3.0.2
keyring==24.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.15
packaging==23.2
pkginfo==1.9.6
Pygments==2.17.2
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyuwsgi==2.0.23
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
six==1.16.0
SQLAlchemy==2.0.24
sqlparse==0.4.4
tomli==2.0.1
twine==4.0.2
typing_extensions==4.9.0
waitress==2.1.2
Werkzeug==3.0.1
zipp==3.17.0
zope.interface==6.1
urllib3